+++ router/router.go
@@ -1,7 +1,7 @@
 package router
 
 import (
-	"github.com/yunzhubanban/server/app/api/endpoint/service"
+	EndpointService "github.com/yunzhubanban/server/app/api/endpoint/service"
 
 	"github.com/gogf/gf/frame/g"
 	"github.com/gogf/gf/net/ghttp"
@@ -13,8 +13,8 @@ func init() {
 		group.Group("/endpoint", func(group *ghttp.RouterGroup) {
 			group.Group("/service", func(group *ghttp.RouterGroup) {
 				group.
-					GET("status", service.Status).
-					PUT("devices", service.Devices)
+					GET("status", EndpointService.Status).
+					PUT("devices", EndpointService.Devices)
 			})
 		})
 	})
